# mypostbank-ui/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI specification - Yasin",
    "description": "OpenAPI Documentation for MyPostBank",
    "termsOfService": "Terms of Service",
    "contact": {
      "name": "Yasin",
      "email": "yasinjb002@gmail.com"
    },
    "license": {
      "name": "License"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/mypostbank",
      "description": "Local ENV"
    },
    {
      "url": "https://yasin",
      "description": "PROV ENV"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/user/accounts/{accountId}/activate": {
      "put": {
        "tags": [
          "bank-account-controller"
        ],
        "summary": "Activate account by ID",
        "operationId": "activateAccount",
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "description": "Account ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/send-reset-email": {
      "post": {
        "tags": [
          "authentication-controller"
        ],
        "summary": "Send password reset email",
        "operationId": "sendResetEmail",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/reset-password": {
      "post": {
        "tags": [
          "authentication-controller"
        ],
        "summary": "Reset password",
        "operationId": "resetPassword",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "newPassword",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/register": {
      "post": {
        "tags": [
          "authentication-controller"
        ],
        "summary": "Register a new user",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/authenticate": {
      "post": {
        "tags": [
          "authentication-controller"
        ],
        "summary": "Authenticate user",
        "operationId": "authenticate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthenticationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/accounts/{accountId}/activate": {
      "post": {
        "tags": [
          "admin-controller"
        ],
        "summary": "Activate account by ID",
        "operationId": "activateAccount_1",
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "description": "Account ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/activate-account": {
      "get": {
        "tags": [
          "authentication-controller"
        ],
        "summary": "Activate user account",
        "operationId": "confirm",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "description": "Activation token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RegisterRequest": {
        "required": [
          "email",
          "firstName",
          "lastName",
          "password",
          "phone"
        ],
        "type": "object",
        "properties": {
          "firstName": {
            "pattern": "^[a-zA-Z]+$",
            "type": "string"
          },
          "lastName": {
            "pattern": "^[a-zA-Z]+$",
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "pattern": "^[0-9]+$",
            "type": "string"
          },
          "password": {
            "maxLength": 20,
            "minLength": 8,
            "pattern": "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[/@$!%*?&])[A-Za-z\\d/@$!%*?&]+$",
            "type": "string"
          }
        }
      },
      "AuthenticationResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          }
        }
      },
      "AuthenticationRequest": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT auth",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}